feat: extract card rendering to separate modules
