+++ console.html
@@ -5,6 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>ADB Console - JTech MDM Installer</title>
     <link rel="stylesheet" href="css/styles.css">
+    <link rel="stylesheet" href="css/cards.css">
     <link rel="preconnect" href="https://fonts.googleapis.com">
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
     <link href="https://fonts.googleapis.com/css2?family=Inter:wght@300;400;500;600;700&display=swap" rel="stylesheet">
@@ -185,6 +186,6 @@
     </div>
 
     <script src="js/webadb.js"></script>
-    <script src="js/app.js" type="module"></script>
+    <script src="js/apps.js" type="module"></script>
 </body>
 </html>
