+++ template.html
@@ -6,6 +6,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>JTech MDM Installer - Android Device Management</title>
     <link rel="stylesheet" href="css/styles.css">
+    <link rel="stylesheet" href="css/cards.css">
     <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/swiper@11/swiper-bundle.min.css" />
     <link rel="preconnect" href="https://fonts.googleapis.com">
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
@@ -182,6 +183,6 @@
 
     <script src="js/webadb.js"></script>
     <script src="https://cdn.jsdelivr.net/npm/swiper@11/swiper-bundle.min.js"></script>
-    <script src="js/app.js" type="module"></script>
+    <script src="js/apps.js" type="module"></script>
 </body>
 </html>
